feat(observatory): use with_ pattern and move modules for consistency
