+++ src/observatory/src/memory/manager.rs
@@ -1,5 +1,6 @@
+use crate::memory::state::services::mutate_runtime_state;
 use crate::types::runtime::RuntimeState;
-use crate::types::state::{StableState, State};
+use crate::types::state::StableState;
 use ic_stable_structures::memory_manager::{MemoryId, MemoryManager};
 use ic_stable_structures::{DefaultMemoryImpl, StableBTreeMap};
 use junobuild_shared::types::memory::Memory;
@@ -9,16 +10,12 @@ const UPGRADES: MemoryId = MemoryId::new(0);
 const NOTIFICATIONS: MemoryId = MemoryId::new(1);
 
 thread_local! {
-    pub static STATE: RefCell<State> = RefCell::default();
-
-    pub static RUNTIME_STATE: RefCell<RuntimeState> = RefCell::default();
-
     static MEMORY_MANAGER: RefCell<MemoryManager<DefaultMemoryImpl>> =
         RefCell::new(MemoryManager::init(DefaultMemoryImpl::default()));
 }
 
 pub fn init_runtime_state() {
-    RUNTIME_STATE.with(|state| *state.borrow_mut() = RuntimeState::default());
+    mutate_runtime_state(|runtime| *runtime = RuntimeState::default());
 }
 
 pub fn get_memory_upgrades() -> Memory {
