+++ src/observatory/src/memory/mod.rs
@@ -1,2 +1,5 @@
 pub mod lifecycle;
-pub mod manager;
+mod manager;
+pub mod state;
+
+pub use manager::init_stable_state;
