+++ src/observatory/src/store/mod.rs
@@ -1,4 +1,3 @@
 mod filter;
 pub mod heap;
-mod services;
 pub mod stable;
